Se agrega cambios

+++ smartbeat_frontend/lib/home/forms/presion_form.dart
@@ -2,23 +2,23 @@ import 'package:reactive_forms/reactive_forms.dart';
 import 'package:smartbeat_frontend/shared/extensions/form_group_extension.dart';
 
 class PresionForm extends FormGroup {
-  FormControl<double> get bloodPressureSistolic =>
-      get<double>('bloodPressureSistolic');
+  FormControl<int> get bloodPressureSistolic =>
+      get<int>('bloodPressureSistolic');
 
-  FormControl<double> get bloodPressureDiastolic =>
-      get<double>('bloodPressureDiastolic');
+  FormControl<int> get bloodPressureDiastolic =>
+      get<int>('bloodPressureDiastolic');
 
   FormControl<int> get bloodPresheartRatesureDiastolic =>
       get<int>('bloodPresheartRatesureDiastolic');
 
   PresionForm()
       : super(fb.group({
-          'bloodPressureSistolic': FormControl<double>(
+          'bloodPressureSistolic': FormControl<int>(
             validators: [
               Validators.required,
             ],
           ),
-          'bloodPressureDiastolic': FormControl<double>(
+          'bloodPressureDiastolic': FormControl<int>(
             validators: [
               Validators.required,
             ],
